+++ backend-spring/leaper/src/main/java/com/ssafy/leaper/domain/chat/service/ChatService.java
@@ -13,30 +13,30 @@ public interface ChatService {
     /**
      * 채팅방 생성
      */
-    ServiceResult<ChatRoomCreateResponse> createChatRoom(Long influencerId, Long advertiserId);
+    ServiceResult<ChatRoomCreateResponse> createChatRoom(Integer influencerId, Integer advertiserId);
 
     /**
      * 채팅방 목록 조회
      */
-    ServiceResult<ChatRoomListResponse> getChatRoomList(Long currentUserId, String userRole);
+    ServiceResult<ChatRoomListResponse> getChatRoomList(Integer currentUserId, String userRole);
 
     /**
      * 채팅 메시지 조회
      */
-    ServiceResult<ChatMessageListResponse> getChatMessages(Long chatRoomId, String before, String after, int size);
+    ServiceResult<ChatMessageListResponse> getChatMessages(Integer chatRoomId, String before, String after, int size);
 
     /**
      * 텍스트 메시지 전송
      */
-    ServiceResult<Void> sendTextMessage(Long chatRoomId, ChatMessageSendRequest request);
+    ServiceResult<Void> sendTextMessage(Integer chatRoomId, ChatMessageSendRequest request);
 
     /**
      * 파일/이미지 메시지 전송 (파일 업로드만 처리, 다운로드 URL 반환)
      */
-    ServiceResult<String> sendFileMessage(Long chatRoomId, Long senderId, String userRole, String messageType, MultipartFile file);
+    ServiceResult<String> sendFileMessage(Integer chatRoomId, Integer senderId, String userRole, String messageType, MultipartFile file);
 
     /**
      * 채팅방 나가기
      */
-    ServiceResult<Void> leaveChatRoom(Long chatRoomId, Long currentUserId, String userRole);
+    ServiceResult<Void> leaveChatRoom(Integer chatRoomId, Integer currentUserId, String userRole);
 }
